# public/swagger.yaml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: QuiMangeOù API
  description: "API pour l'application QuiMangeOù : https://github.com/VivienZo/QuiMangeOuApi "
# the domain of the service
host: localhost:8080
...
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
produces:
  - application/json
paths:
  /validations:
    post:
      summary: "Permet de demander la création d'un nouveau Groupe. La requête prend un e-mail et un nom de groupe en entrée. Un document Validation est créé et un token de sécurité lui est associé. Ce token servira à valider la création du nouveau Groupe. Un e-mail est envoyé à l'adresse fournie pour que l'utilisateur puisse valider son nouveau Groupe. La réponse HTTP contient un lien qui permet de valider la création du groupe."
      parameters:
        - name: nomGroupe
          in: body
          description: "Nom du groupe"
          required: true
          type: string
        - name: email
          in: body
          description: "E-mail de contact du groupe"
          required: true
          type: string
      tags:
        - Validations
      responses:
        201:
          description: "OK : {status, data, message}"
        500:
          description: "KO : {status, error, message}"
  /validations/:token:
    get:
      summary: "Vérifie si un document Validation contient le token fourni et si celui-ci n'est pas expiré. Si ces deux conditions sont respectées, le Groupe rattaché à la Validation est créé. Un e-mail contenant un lien vers le Groupe est envoyé à l'adresse de contact. Ce lien peut être partagé par l'utilisateur pour inviter d'autres personnes à participer à son groupe."
      parameters:
        - name: token
          in: path
          description: "Clé de validation du groupe"
          required: true
          type: string
      tags:
        - Validations
      responses:
        200:
          description: "OK : {status, data, message}"
        500:
          description: "KO : {status, error, message}"
...